+++ c/main-hvapprox.c
@@ -1,13 +1,11 @@
 /*************************************************************************
 
- hv: main program
+ hvapprox: main program
 
  ---------------------------------------------------------------------
 
-                       Copyright (c) 2010
-                  Carlos M. Fonseca <cmfonsec@dei.uc.pt>
+                       Copyright (c) 2025
              Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>
-                    Luis Paquete <paquete@dei.uc.pt>
 
  This program is free software (software libre); you can redistribute
  it and/or modify it under the terms of the GNU General Public License
@@ -31,16 +29,6 @@
 
  Relevant literature:
 
- [1]  C. M. Fonseca, L. Paquete, and M. Lopez-Ibanez. An
-      improved dimension-sweep algorithm for the hypervolume
-      indicator. In IEEE Congress on Evolutionary Computation,
-      pages 1157-1163, Vancouver, Canada, July 2006.
-
- [2]  Nicola Beume, Carlos M. Fonseca, Manuel Lopez-Ibanez, Luis Paquete, and
-      J. Vahrenhold.  On the complexity of computing the hypervolume
-      indicator. IEEE Transactions on Evolutionary Computation,
-      13(5):1075-1082, 2009.
-
 *************************************************************************/
 #include "config.h"
 #include <errno.h>
@@ -50,18 +38,21 @@
 #include <string.h>
 #include <unistd.h>  // for getopt()
 #include <getopt.h> // for getopt_long()
+#include <time.h> // time()
+#include <limits.h> // LONG_MAX
+#include <inttypes.h> // PRIu32
 
-#include "hv.h"
+#include "hvapprox.h"
 #include "timer.h"
-#define CMDLINE_COPYRIGHT_YEARS "2010-2023"
-#define CMDLINE_AUTHORS "Carlos M. Fonseca <cmfonsec@dei.uc.pt>\n" \
-    "Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>\n" \
-    "Luis Paquete <paquete@dei.uc.pt>\n"
+#define CMDLINE_COPYRIGHT_YEARS "2025"
+#define CMDLINE_AUTHORS "Manuel Lopez-Ibanez <manuel.lopez-ibanez@manchester.ac.uk>\n"
 #include "cmdline.h"
 
 static int verbose_flag = 1;
 static bool union_flag = false;
 static char *suffix = NULL;
+
+enum approx_method_t { DZ2019_MC=1, DZ2019_HW=2 };
 
 static void usage(void)
 {
@@ -69,8 +60,9 @@ static void usage(void)
            "Usage: %s [OPTIONS] [FILE...]\n\n", program_invocation_short_name);
 
     printf(
-"Calculate the hypervolume of each input set of each FILE. \n"
-"With no FILE, or when FILE is -, read standard input.\n\n"
+"Approximate the hypervolume value of each input set of each FILE. \n"
+"The approximation uses Monte-Carlo sampling, thus gets more accurate with larger\n"
+"values of --nsamples. With no FILE, or when FILE is -, read standard input.\n\n"
 
 "Options:\n"
 OPTION_HELP_STR
@@ -85,6 +77,11 @@ OPTION_VERSION_STR
 " -s, --suffix=STRING Create an output file for each input file by appending\n"
 "                     this suffix. This is ignored when reading from stdin. \n"
 "                     If missing, output is sent to stdout.                 \n"
+" -n, --nsamples=N    Number of Monte-Carlo samples (N is a positive integer).\n"
+" -m, --method=M      1: Monte-Carlo sampling using normal distribution;    \n"
+"                     2: Hua-Wang deterministic sampling (default).         \n"
+" -S, --seed=S        Seed of the random number generator (S: positive integer).\n"
+"                     Only method=1.                                        \n"
 "\n");
 }
 
@@ -101,9 +98,11 @@ OPTION_VERSION_STR
 
 */
 
+// FIXME: There is a similar function in main-hv.c
 static void
-hv_file (const char *filename, double *reference,
-         double *maximum, double *minimum, int *nobj_p)
+hvapprox_file (const char *filename, double *reference,
+               double *maximum, double *minimum, int *nobj_p,
+               uint_fast32_t nsamples, enum approx_method_t hv_approx_method, uint32_t seed)
 {
     double *data = NULL;
     int *cumsizes = NULL;
@@ -177,10 +176,22 @@ hv_file (const char *filename, double *reference,
         printf ("\n");
     }
 
+    // Minimise everything by default.
+    const bool * maximise = new_bool_maximise(nobj, false);
     for (n = 0, cumsize = 0; n < nruns; cumsize = cumsizes[n], n++) {
         Timer_start ();
 
-        double volume = fpli_hv (&data[nobj * cumsize], nobj, cumsizes[n] - cumsize, reference);
+        double volume;
+        switch (hv_approx_method) {
+          case DZ2019_MC:
+              volume = hv_approx_normal(&data[nobj * cumsize], nobj, cumsizes[n] - cumsize, reference, maximise, nsamples, seed);
+              break;
+          case DZ2019_HW:
+              volume = hv_approx_hua_wang(&data[nobj * cumsize], nobj, cumsizes[n] - cumsize, reference, maximise, nsamples);
+              break;
+          default:
+              unreachable();
+        }
 
         if (volume == 0.0) {
             errprintf ("none of the points strictly dominates the reference point\n");
@@ -200,6 +211,7 @@ hv_file (const char *filename, double *reference,
         fclose (outfile);
         free (outfilename);
     }
+    free ((void *) maximise);
     free (data);
     free (cumsizes);
     if (needs_minimum) {
@@ -212,7 +224,7 @@ hv_file (const char *filename, double *reference,
 int main(int argc, char *argv[])
 {
     /* See the man page for getopt_long for an explanation of these fields.  */
-    static const char short_options[] = "hVvqur:s:S";
+    static const char short_options[] = "hVvqur:s:n:m:S:";
     static const struct option long_options[] = {
         {"help",       no_argument,       NULL, 'h'},
         {"version",    no_argument,       NULL, 'V'},
@@ -221,6 +233,9 @@ int main(int argc, char *argv[])
         {"reference",  required_argument, NULL, 'r'},
         {"union",      no_argument,       NULL, 'u'},
         {"suffix",     required_argument, NULL, 's'},
+        {"method",     required_argument, NULL, 'm'},
+        {"nsamples",   required_argument, NULL, 'n'},
+        {"seed",       required_argument, NULL, 'S'},
         {NULL, 0, NULL, 0} /* marks end of list */
     };
 
@@ -228,6 +243,9 @@ int main(int argc, char *argv[])
 
     double *reference = NULL;
     int nobj = 0;
+    uint32_t seed = 0;
+    uint_fast32_t nsamples = 0;
+    enum approx_method_t hv_approx_method = DZ2019_HW;
 
     int opt; /* it's actually going to hold a char.  */
     int longopt_index;
@@ -250,24 +268,74 @@ int main(int argc, char *argv[])
               suffix = optarg;
               break;
 
-          case 'q': // --quiet
-              verbose_flag = 0;
+          case 'n': { // --nsamples
+              char *endp;
+              long int value = strtol(optarg, &endp, 10);
+              if (endp == optarg || *endp != '\0' || value <= 0 || value == LONG_MAX) {
+                  errprintf ("value of --nsamples must be a positive integer '%s'", optarg);
+                  exit (EXIT_FAILURE);
+              }
+              nsamples = (uint_fast32_t) value;
+              break;
+          }
+
+          case 'm': // --method
+              switch (*optarg) {
+                case '1':
+                    hv_approx_method = DZ2019_MC; break;
+                case '2':
+                    hv_approx_method = DZ2019_HW; break;
+                default:
+                    errprintf ("valid values of --method (-m) are: 1 or 2, not '%s'", optarg);
+                  exit(EXIT_FAILURE);
+              }
               break;
 
-          case 'v': // --verbose
-              verbose_flag = 2;
+          case 'S': {// --seed
+              char *endp;
+              long int value = strtol(optarg, &endp, 10);
+              if (endp == optarg || *endp != '\0' || value <= 0) {
+                  errprintf ("value of --seed must be a positive integer '%s'", optarg);
+                  exit (EXIT_FAILURE);
+              }
+              seed = (uint32_t) value;
               break;
+          }
+        case 'q': // --quiet
+            verbose_flag = 0;
+            break;
 
-          default:
-              default_cmdline_handler(opt);
+        case 'v': // --verbose
+            verbose_flag = 2;
+            break;
+
+        default:
+            default_cmdline_handler(opt);
         }
     }
 
+    if (nsamples == 0) {
+        errprintf ("must specify a value for --nsamples");
+        exit(EXIT_FAILURE);
+    }
+
+    if (seed == 0) {
+        if (hv_approx_method == DZ2019_MC)
+            seed = (uint32_t) time(NULL);
+    } else if (hv_approx_method == DZ2019_HW) {
+        errprintf("cannot use --seed with --method=2");
+        exit(EXIT_FAILURE);
+    }
+
+    if (verbose_flag >= 2)
+        printf("# seed: %"PRIu32 "\n# nsamples: %lu\n", seed, (unsigned long) nsamples);
+
     int numfiles = argc - optind;
     if (numfiles < 1) /* Read stdin.  */
-        hv_file (NULL, reference, NULL, NULL, &nobj);
+        hvapprox_file(NULL, reference, NULL, NULL, &nobj, nsamples, hv_approx_method, seed);
+
     else if (numfiles == 1) {
-        hv_file (argv[optind], reference, NULL, NULL, &nobj);
+        hvapprox_file (argv[optind], reference, NULL, NULL, &nobj, nsamples, hv_approx_method, seed);
     } else {
         int k;
         double *maximum = NULL;
@@ -288,7 +356,7 @@ int main(int argc, char *argv[])
             }
         }
         for (k = 0; k < numfiles; k++)
-            hv_file (argv[optind + k], reference, maximum, minimum, &nobj);
+            hvapprox_file (argv[optind + k], reference, maximum, minimum, &nobj, nsamples, hv_approx_method, seed);
 
         free(minimum);
         free(maximum);