+++ r/NEWS.md
@@ -1,7 +1,7 @@
 # moocore 0.1.7.900
 
  * Document the EAF and Vorob'ev expectation and deviation in more detail.
-
+ * New function `hv_approx()`.
 
 # moocore 0.1.7